Update spec repo fixture to latest 0.6 and fix CP version checking.

+++ lib/cocoapods/command/setup.rb
@@ -89,9 +89,12 @@ module Pod
         else
           add_master_repo_command.run
         end
-        hook = config.repos_dir + 'master/.git/hooks/pre-commit'
-        hook.open('w') { |f| f << "#!/bin/sh\nrake lint" }
-        `chmod +x '#{hook}'`
+        # Mainly so the specs run with submodule repos
+        if (dir + '.git/hooks').exist?
+          hook = dir + '.git/hooks/pre-commit'
+          hook.open('w') { |f| f << "#!/bin/sh\nrake lint" }
+          `chmod +x '#{hook}'`
+        end
         puts "Setup completed (#{push? ? "push" : "read-only"} access)" unless config.silent
       end
     end
